requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
airportsdata==20250523
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
asttokens @ file:///croot/asttokens_1743630435401/work
attrs==25.3.0
backports.tarfile==1.2.0
blake3==1.0.5
blessed==1.21.0
cachetools==6.0.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
comm @ file:///croot/comm_1709322850197/work
compressed-tensors==0.9.3
contourpy==1.3.2
cryptography==45.0.5
cupy-cuda12x==13.4.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.6.0
debugpy @ file:///croot/debugpy_1736267418885/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
Deprecated==1.2.18
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docutils==0.21.2
einops==0.8.1
elastic-transport==8.17.1
elasticsearch==8.11.1
email_validator==2.2.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastapi==0.115.12
fastapi-cli==0.0.7
fastrlock==0.8.3
ffmpy==0.6.0
filelock==3.18.0
flashinfer-python==0.2.5
fonttools==4.58.4
frozenlist==1.6.0
fsspec==2025.3.0
gguf==0.17.0
googleapis-common-protos==1.70.0
gpustat==1.1.1
gradio==5.34.2
gradio_client==1.10.3
greenlet==3.2.2
groovy==0.1.2
grpcio==1.71.0
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.3
id==1.5.0
idna==3.10
importlib_metadata==8.0.0
interegular==0.3.3
ipykernel @ file:///croot/ipykernel_1737660677549/work
ipython @ file:///croot/ipython_1744912971218/work
ipython_pygments_lexers @ file:///croot/ipython_pygments_lexers_1744753235686/work
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jedi @ file:///croot/jedi_1733987392413/work
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client @ file:///croot/jupyter_client_1737570961872/work
jupyter_core @ file:///croot/jupyter_core_1718818295206/work
keyring==25.6.0
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-text-splitters==0.3.8
langsmith==0.3.43
lark==1.2.2
llguidance==0.7.26
llvmlite==0.44.0
lm-format-enforcer==0.10.11
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline @ file:///work/ci_py311/matplotlib-inline_1676823841154/work
mdurl==0.1.2
mistral_common==1.5.6
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx==3.5
nh3==0.2.21
ninja==1.11.1.4
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.82.1
opencv-python-headless==4.11.0.86
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.9
orjson==3.10.18
outlines==0.1.11
outlines_core==0.1.26
packaging @ file:///croot/packaging_1734472117206/work
pandas==2.2.3
parso @ file:///croot/parso_1733963305961/work
partial-json-parser==0.2.1.1.post5
pexpect @ file:///croot/pexpect_1750072593878/work
pillow==11.2.1
pip @ file:///croot/pip_1746204010231/work
pipdeptree==2.26.1
platformdirs @ file:///croot/platformdirs_1744273042065/work
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.0
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
propcache==0.3.1
protobuf==4.25.8
psutil==7.0.0
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments @ file:///croot/pygments_1744664109463/work
pyparsing==3.2.3
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
ray==2.46.0
readme_renderer==44.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.25.1
ruff==0.12.1
safehttpx==0.1.6
safetensors==0.5.3
scipy==1.15.3
seaborn==0.13.2
SecretStorage==3.3.3
semantic-version==2.10.0
sentencepiece==0.2.0
setuptools==80.9.0
shellingham==1.5.4
six @ file:///croot/six_1744271502820/work
sniffio==1.3.1
SQLAlchemy==2.0.41
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.46.2
sympy==1.13.1
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tomlkit==0.13.3
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado @ file:///croot/tornado_1748956929273/work
tqdm==4.67.1
traitlets @ file:///croot/traitlets_1718227057033/work
transformers==4.52.4
triton==3.2.0
twine==6.1.0
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
vllm==0.8.5
watchfiles==1.0.5
wcwidth @ file:///croot/wcwidth_1750352883074/work
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
xformers==0.0.29.post2
xgrammar==0.1.18
xxhash==3.5.0
yarl==1.20.0
zipp==3.22.0
zstandard==0.23.0
